+++ .gitignore
@@ -5,3 +5,4 @@
 /utils/__pycache__/
 /colors.txt
 /utils/helpstrings.py/
+/.theia
