+++ README.md
@@ -7,7 +7,7 @@
 
 - If you find any issues within the code please submit a issue via this Repo's [Issue Page](https://github.com/Yat-o/HimejiBot/issues)
 
-- HimejiBot Uses Python Version [3.8.6](https://www.python.org/downloads/release/python-386/) and [Discord.py](https://discordpy.readthedocs.io/en/latest/#) Version 1.5.1
+- HimejiBot Uses Python Version [3.8.6](https://www.python.org/downloads/release/python-386/) and [Discord.py](https://discordpy.readthedocs.io/en/latest/#) Version 1.6.0a
 
 # TODO
 - C&C Cogs and perhaps bot.py
