+++ cogs/misc.py
@@ -1,44 +1,15 @@
 import discord
-import aiohttp 
 
 from discord.ext import commands
 
 class Misc(commands.Cog):
-    
+
     def __init__(self, bot):
         self.bot = bot
-
-    @commands.command()
-    async def megumin(self, ctx):
-        async with aiohttp.ClientSession() as cs:
-            async with cs.get('https://waifu.pics/api/sfw/megumin') as r:
-
-                meguminpic = (await r.json())['url']
-
-                await ctx.send(meguminpic)
-
-
-    @commands.command()
-    async def shinobu(self, ctx):
-        async with aiohttp.ClientSession() as cs:
-            async with cs.get('https://waifu.pics/api/sfw/shinobu') as r:
-                
-                shinobupic = (await r.json())['url']
-
-                await ctx.send(shinobupic)
-
-    @commands.command()
-    async def neko(self, ctx):
-        async with aiohttp.ClientSession() as cs:
-            async with cs.get('https://waifu.pics/api/sfw/neko'):
-
-                nekopic = (await r.json())['url']
-
-                await ctx.send(nekopic)
-
- 
+    
+    @commands.command(aliases = ['issue'])
+    async def bug(self, ctx):
+        await ctx.send('See a bug or issue with the bot. Please report it at https://github.com/Yat-o/HimejiBot/issues')
+    
 def setup(bot):
     bot.add_cog(Misc(bot))
-    
-
-
